# Admin/bin/Debug/netcoreapp3.1/Docs/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Admin API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sim-groups/mine": {
      "post": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_PostSimMine",
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSimGroupParam"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "get": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_GetSimGroups",
        "parameters": [
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortKey10"
            },
            "x-position": 1
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Order"
            },
            "x-position": 2
          },
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "Apn",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000.0,
              "minimum": 1.0
            },
            "x-position": 6
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfSimGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
    "/api/sim-groups": {
      "post": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_PostSim",
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSimGroupAdminParam"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      },
      "get": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_GetSimGroups2",
        "parameters": [
          {
            "name": "OrganizationCode",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortKey10"
            },
            "x-position": 2
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Order"
            },
            "x-position": 3
          },
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "Apn",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000.0,
              "minimum": 1.0
            },
            "x-position": 7
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfSimGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      }
    },
    "/api/sim-groups/mine/{id}": {
      "delete": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_DeleteSimGroupMine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "get": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_GetSimGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_UpdateSimGroupMine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutSimGroupParam"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
    "/api/sim-groups/{id}": {
      "delete": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_DeleteSimGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      },
      "get": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_GetSimGroup2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "SimGroups"
        ],
        "operationId": "SimGroups_UpdateSimGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutSimGroupParam"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      }
    },
    "/api/sims/mine": {
      "post": {
        "tags": [
          "Sims"
        ],
        "operationId": "Sims_PostSimMine",
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterParam3"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sim"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "get": {
        "tags": [
          "Sims"
        ],
        "operationId": "Sims_GetSims",
        "parameters": [
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortKey11"
            },
            "x-position": 1
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Order"
            },
            "x-position": 2
          },
          {
            "name": "UserName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "Msisdn",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "SimGroupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000.0,
              "minimum": 1.0
            },
            "x-position": 7
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfSim"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
...
    "/api/sims/mine/{id}": {
      "delete": {
        "tags": [
          "Sims"
        ],
        "operationId": "Sims_DeleteSimMine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sim"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "get": {
        "tags": [
          "Sims"
        ],
        "operationId": "Sims_GetSim",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sim"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Sims"
        ],
        "operationId": "Sims_PutSimMine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterParam3"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SimGroup": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "apn",
          "nasIp",
          "isolatedNw1IpPool",
          "isolatedNw1IpRange",
          "authenticationServerIp",
          "primaryDns",
          "secondaryDns",
          "userNameSuffix"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "organizationCode": {
            "type": "integer",
            "format": "int32"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "apn": {
            "type": "string",
            "minLength": 1
          },
          "nasIp": {
            "type": "string",
            "minLength": 1
          },
          "isolatedNw1IpPool": {
            "type": "string",
            "minLength": 1
          },
          "isolatedNw1IpRange": {
            "type": "string",
            "minLength": 1
          },
          "authenticationServerIp": {
            "type": "string",
            "minLength": 1
          },
          "primaryDns": {
            "type": "string",
            "minLength": 1
          },
          "secondaryDns": {
            "type": "string",
            "minLength": 1
          },
          "isolatedNw1PrimaryDns": {
            "type": "string",
            "nullable": true
          },
          "isolatedNw1SecondaryDns": {
            "type": "string",
            "nullable": true
          },
          "userNameSuffix": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "Order": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Asc",
          "Desc"
        ],
        "enum": [
          0,
          1
        ]
      },
...
      "PostSimGroupParam": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SimParam"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
...
      "SortKey10": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Name",
          "Apn",
          "AuthenticationServerIp",
          "NasIp"
        ],
        "enum": [
          0,
          1,
          2,
          3
        ]
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ],
...
}